[logic][xl]: db_logic moved into logic __init__. tests updated, moks updated, imports updated - refs #376
