+++ app/__init__.py
@@ -18,13 +18,12 @@ from flask_oauthlib.client import OAuth
 from werkzeug.utils import import_string
 from werkzeug.exceptions import NotFound, Unauthorized, MethodNotAllowed, BadRequest
 from .database import db
-from .logic import ma, db_logic
+from .logic import ma, User
 from app.auth.controllers import auth_blueprint, bitstore_blueprint
 from app.auth.jwt import JWT
 from app.package.controllers import package_blueprint
 from app.site.controllers import site_blueprint
 from app.profile.controllers import profile_blueprint
-from app.profile.models import User
 from app.search.controllers import search_blueprint
 from app.utils import InvalidUsage
 from flask import jsonify
@@ -148,6 +147,6 @@ def create_app():
         g.current_user = None
         if token:
             payload = JWT(app.config['JWT_SEED']).decode(token)
-            g.current_user = db_logic.User.get(payload['user'])
+            g.current_user = User.get(payload['user'])
 
     return app
