+++ app/logic/__init__.py
@@ -5,6 +5,7 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 
 import json
+import os
 
 from BeautifulSoup import BeautifulSoup
 from flask import request, session
@@ -13,15 +14,271 @@ from sqlalchemy.orm.exc import NoResultFound
 
 from app.auth.annotations import check_is_authorized, get_user_from_jwt
 from app.auth.jwt import JWT, FileData
+from app.database import db
 from app.bitstore import BitStore
-from app.package.models import Package, PackageStateEnum
-from app.profile.models import Publisher, User
-from app.logic import db_logic
 from app.logic.search import DataPackageQuery
 from app.utils import InvalidUsage
 from app.utils.helpers import text_to_markdown, dp_in_readme
+import app.models as models
 
-from .db_logic import ma
+
+from flask_marshmallow import Marshmallow
+from marshmallow import pre_load, pre_dump
+from marshmallow_enum import EnumField
+
+
+ma = Marshmallow()
+
+
+class LogicBase(object):
+
+    schema = None
+
+    @classmethod
+    def serialize(cls, sqla_instance):
+        if sqla_instance is None:
+            return None
+        serialized = cls.schema().dump(sqla_instance).data
+        return serialized
+
+    @classmethod
+    def deserialize(cls, dict_object):
+        deserialized = cls.schema().load(dict_object, session=db.session).data
+        return deserialized
+
+
+#####################################################
+# Packages
+
+class PackageSchema(ma.ModelSchema):
+    class Meta:
+        model = models.Package
+
+    status = EnumField(models.PackageStateEnum)
+
+
+class PackageTagSchema(ma.ModelSchema):
+    class Meta:
+        model = models.PackageTag
+
+
+class PackageMetadataSchema(ma.Schema):
+    class Meta:
+        fields = ('id', 'name', 'publisher', 'readme', 'descriptor')
+
+    publisher = ma.Method('get_publisher_name')
+    readme = ma.Method('get_readme')
+    descriptor = ma.Method('get_descriptor')
+
+    def get_publisher_name(self, data):
+        return data.publisher.name
+
+    def get_readme(self, data):
+        version = filter(lambda t: t.tag == 'latest', data.tags)[0]
+        return version.readme or ''
+
+    def get_descriptor(self, data):
+        version = filter(lambda t: t.tag == 'latest', data.tags)[0]
+        return version.descriptor
+
+
+class Package(LogicBase):
+    schema = PackageMetadataSchema
+
+    @classmethod
+    def get(cls, publisher, package):
+        data = models.Package.get_by_publisher(publisher, package)
+        return cls.serialize(data)
+
+    @classmethod
+    def exists(cls, publisher, package):
+        instance = models.Package.get_by_publisher(publisher, package)
+        return instance is not None
+
+    @classmethod
+    def delete(cls, publisher, package):
+        pkg = models.Package.get_by_publisher(publisher, package)
+        # TODO: should be able to db.session.delete(pkg) but deletes publishers!
+        models.Package.query.filter(models.Package.id == pkg.id).delete()
+        db.session.commit()
+        return True
+
+    @classmethod
+    def create_or_update_tag(cls, publisher, package, tag):
+        package = models.Package.get_by_publisher(publisher, package)
+
+        data_latest = models.PackageTag.query.join(models.Package)\
+            .filter(models.Package.id == package.id,
+                    models.PackageTag.tag == 'latest').one()
+
+        tag_instance = models.PackageTag.query.join(models.Package) \
+            .filter(models.Package.id == package.id,
+                    models.PackageTag.tag == tag).first()
+
+        update_props = ['descriptor', 'readme', 'package_id']
+        if tag_instance is None:
+            tag_instance = models.PackageTag()
+
+        for update_prop in update_props:
+            setattr(tag_instance, update_prop, getattr(data_latest, update_prop))
+        tag_instance.tag = tag
+
+        db.session.add(tag_instance)
+        db.session.commit()
+        return True
+
+    @classmethod
+    def create_or_update(cls, name, publisher_name, **kwargs):
+        pub_id = models.Publisher.query.filter_by(name=publisher_name).one().id
+        instance = models.Package.get_by_publisher(publisher_name, name)
+
+        if instance is None:
+            instance = models.Package(name=name)
+            instance.publisher_id = pub_id
+            tag_instance = models.PackageTag()
+            instance.tags.append(tag_instance)
+        else:
+            tag_instance = models.PackageTag.query.join(models.Package) \
+                .filter(models.Package.id == instance.id,
+                        models.PackageTag.tag == 'latest').one()
+        for key, value in kwargs.items():
+            if key not in ['descriptor', 'readme']:
+                setattr(instance, key, value)
+            else:
+                setattr(tag_instance, key, value)
+        db.session.add(instance)
+        db.session.commit()
+
+    @classmethod
+    def change_status(cls, publisher_name,
+                            package_name, status=models.PackageStateEnum.active):
+        pkg = models.Package.get_by_publisher(publisher_name, package_name)
+        pkg.status = status
+        db.session.add(pkg)
+        db.session.commit()
+        return True
+
+
+class PackageTag(LogicBase):
+    schema = PackageTagSchema
+
+    @classmethod
+    def get(cls, package_id, tag):
+        tag = models.PackageTag.get_by_tag(package_id, tag)
+        return cls.serialize(tag)
+
+#####################################################
+# Profiles - Publishers and Users
+
+class PublisherSchema(ma.ModelSchema):
+    class Meta:
+        model = models.Publisher
+        dateformat = ("%B %Y")
+
+    contact = ma.Method('add_public_contact')
+    joined = ma.DateTime(attribute = 'created_at')
+
+    def add_public_contact(self, data):
+        if data.contact_public:
+            contact = dict(phone=data.phone,
+                           email=data.email,
+                           country=data.country)
+            return contact
+
+
+class UserSchema(ma.ModelSchema):
+    class Meta:
+        model = models.User
+
+
+class UserInfoSchema(ma.Schema):
+    class Meta:
+        fields = ('email', 'login', 'name')
+
+    @pre_load
+    def load_user(self, data):
+        email = data.get('email')
+        emails = data.get('emails')
+
+        if email:
+            return data
+
+        if not emails or not len(emails):
+            raise InvalidUsage('Email Not Found', 404)
+
+        for email in emails:
+            if email.get('primary') == 'true':
+                data['email'] = email.get('email')
+                return data
+
+
+class PublisherUserSchema(ma.ModelSchema):
+    class Meta:
+        model = models.PublisherUser
+
+    role = EnumField(models.UserRoleEnum)
+
+
+class Publisher(LogicBase):
+
+    schema = PublisherSchema
+
+    @classmethod
+    def get(cls, publisher):
+        pub = models.Publisher.get_by_name(publisher)
+        return cls.serialize(pub)
+
+    @classmethod
+    def create(cls, metadata):
+        pub = cls.deserialize(metadata)
+        db.session.add(pub)
+        db.session.commit()
+        return pub
+
+class User(LogicBase):
+
+    schema = UserSchema
+
+    @classmethod
+    def get(cls, usr_id):
+        usr = models.User.query.get(usr_id)
+        return cls.serialize(usr)
+
+    @classmethod
+    def create(cls, metadata):
+        usr = cls.deserialize(metadata)
+        db.session.add(usr)
+        db.session.commit()
+        return usr
+
+    @classmethod
+    def find_or_create(cls, user_info):
+        """
+        This method populates db when user sign up or login through external auth system
+        :param user_info: User data from external auth system
+        :param oauth_source: From which oauth source the user coming from e.g. github
+        :return: User data from Database
+        """
+        print(user_info)
+        user = models.User.get_by_name(user_info['login'])
+        if user:
+            return user
+
+        # convert github info to our local user info
+        ourinfo = {
+            'name': user_info['login'],
+            'full_name': user_info.get('name'),
+            'email': user_info.get('email')
+        }
+        user = models.User(**ourinfo)
+        # create publisher for this user
+        publisher = models.Publisher(name=user.name)
+        association = models.PublisherUser(role=models.UserRoleEnum.owner, publisher=publisher, user=user)
+        # user.publishers.append(association)
+        db.session.add(user)
+        db.session.commit()
+        return user
+
 
 # TODO: authz
 def get_package(publisher, package):
@@ -29,7 +286,7 @@ def get_package(publisher, package):
     Returns info for package - modified descriptor, bitstore URL for descriptor,
     views and short README
     '''
-    metadata = db_logic.Package.get(publisher, package)
+    metadata = Package.get(publisher, package)
     if not metadata:
         return None
 
@@ -65,7 +322,7 @@ def finalize_package_publish(user_id, datapackage_url):
     Returns status "queued" if ok, else - None
     '''
     publisher, package, version = BitStore.extract_information_from_s3_url(datapackage_url)
-    if db_logic.Package.exists(publisher, package):
+    if Package.exists(publisher, package):
         status = check_is_authorized('Package::Update', publisher, package, user_id)
     else:
         status = check_is_authorized('Package::Create', publisher, package, user_id)
@@ -78,7 +335,7 @@ def finalize_package_publish(user_id, datapackage_url):
     body = json.loads(b)
     bit_store.change_acl('public-read')
     readme = bit_store.get_s3_object(bit_store.get_readme_object_key())
-    db_logic.Package.create_or_update(name=package, publisher_name=publisher,
+    Package.create_or_update(name=package, publisher_name=publisher,
                              descriptor=body, readme=readme)
     return "queued"
 
@@ -98,7 +355,7 @@ def get_authorized_user_info():
     emails = github.get('user/emails').data
     user_info['emails'] = emails
 
-    user_info_schema = db_logic.UserInfoSchema()
+    user_info_schema = UserInfoSchema()
     user_info = user_info_schema.load(user_info).data
 
     return user_info
@@ -118,7 +375,7 @@ def get_jwt_token():
         raise InvalidUsage('Secret can not be empty', 400)
     elif user_name is not None:
         try:
-            user = User.query.filter_by(name=user_name).one()
+            user = models.User.query.filter_by(name=user_name).one()
         except NoResultFound as e:
             app.logger.error(e)
             raise InvalidUsage('user does not exists', 404)
@@ -127,7 +384,7 @@ def get_jwt_token():
             user_id = user.id
     elif email is not None:
         try:
-            user = User.query.filter_by(email=email).one()
+            user = models.User.query.filter_by(email=email).one()
         except NoResultFound as e:
             app.logger.error(e)
             raise InvalidUsage('user does not exists', 404)
@@ -151,7 +408,7 @@ def generate_signed_url():
     publisher, package_name = metadata['owner'], metadata['name']
     res_payload = {'filedata': {}}
 
-    if db_logic.Package.exists(publisher, package_name):
+    if Package.exists(publisher, package_name):
         status = check_is_authorized('Package::Update', publisher, package_name, user_id)
     else:
         status = check_is_authorized('Package::Create', publisher, package_name, user_id)
