+++ tests/auth/test_controller.py
@@ -188,7 +188,7 @@ class CallbackHandlingTestCase(unittest.TestCase):
     @patch('flask_oauthlib.client.OAuthRemoteApp.authorized_response')
     @patch('flask_oauthlib.client.OAuthRemoteApp.get')
     @patch('app.auth.jwt.JWT.encode')
-    @patch('app.logic.db_logic.User.find_or_create')
+    @patch('app.logic.User.find_or_create')
     def test_gets_private_email_and_return_200_if_all_right(self,
                                      create_user, jwt_helper, get_user,
                                      authorized_response):
@@ -218,7 +218,7 @@ class CallbackHandlingTestCase(unittest.TestCase):
     @patch('flask_oauthlib.client.OAuthRemoteApp.authorized_response')
     @patch('flask_oauthlib.client.OAuthRemoteApp.get')
     @patch('app.auth.jwt.JWT.encode')
-    @patch('app.logic.db_logic.User.find_or_create')
+    @patch('app.logic.User.find_or_create')
     def test_return_200_if_all_right(self,
                                      create_user, jwt_helper, get_user,
                                      authorized_response):
