+++ tests/logic/test_logic.py
@@ -6,7 +6,7 @@ from app import create_app
 from app.database import db
 from app.package.models import *
 from app.profile.models import *
-from app.logic import *
+import app.logic as logic
 from app.utils import InvalidUsage
 
 def create_test_package(publisher='demo', package='demo-package', descriptor={}):
@@ -42,7 +42,7 @@ class DataPackageShowTest(unittest.TestCase):
 
     def test_get_package(self):
         # result is a dict ready for passing to templates or API
-        package = get_package(self.publisher, self.package)
+        package = logic.get_package(self.publisher, self.package)
         self.assertEqual(package['descriptor']['name'], self.package)
         self.assertEqual(package['descriptor']['owner'], self.publisher)
         self.assertTrue(package.get('datapackag_url'))
@@ -50,11 +50,11 @@ class DataPackageShowTest(unittest.TestCase):
         self.assertEqual(package['short_readme'], '')
 
     def test_returns_none_if_package_not_found(self):
-        package = get_package(self.publisher, 'unknown')
+        package = logic.get_package(self.publisher, 'unknown')
         self.assertIsNone(package)
-        package = get_package('unknown', self.package)
+        package = logic.get_package('unknown', self.package)
         self.assertIsNone(package)
-        package = get_package('unknown', 'unknown')
+        package = logic.get_package('unknown', 'unknown')
         self.assertIsNone(package)
 
     def tearDown(self):
@@ -82,7 +82,7 @@ class PackageTest(unittest.TestCase):
             create_test_package(self.publisher, self.package, self.descriptor)
 
 
-    @patch('app.logic.db_logic.Package.create_or_update')
+    @patch('app.logic.Package.create_or_update')
     @patch('app.bitstore.BitStore.get_metadata_body')
     @patch('app.bitstore.BitStore.get_readme_object_key')
     @patch('app.bitstore.BitStore.get_s3_object')
@@ -96,11 +96,11 @@ class PackageTest(unittest.TestCase):
         get_readme_object_key.return_value = ''
         get_s3_object.return_value = ''
         change_acl.return_value = None
-        status = finalize_package_publish(1, self.datapackage_url)
+        status = logic.finalize_package_publish(1, self.datapackage_url)
         self.assertEqual(status, 'queued')
 
 
-    @patch('app.logic.db_logic.Package.create_or_update')
+    @patch('app.logic.Package.create_or_update')
     @patch('app.bitstore.BitStore.get_metadata_body')
     @patch('app.bitstore.BitStore.get_readme_object_key')
     @patch('app.bitstore.BitStore.get_s3_object')
@@ -115,7 +115,7 @@ class PackageTest(unittest.TestCase):
         get_s3_object.return_value = ''
         change_acl.return_value = None
         with self.assertRaises(InvalidUsage) as context:
-            finalize_package_publish(2, self.datapackage_url)
+            logic.finalize_package_publish(2, self.datapackage_url)
         self.assertEqual(context.exception.status_code, 400)
 
 
@@ -133,25 +133,25 @@ class HelpersTest(unittest.TestCase):
 
 
     def test_validate_for_jinja_returns_descriptor_if_no_licenses(self):
-        descriptor = validate_for_template(self.descriptor)
+        descriptor = logic.validate_for_template(self.descriptor)
         self.assertEqual(self.descriptor, descriptor)
 
     def test_validate_for_jinja_returns_descriptor_if_licenses_is_list(self):
         self.descriptor['licenses'] = []
-        descriptor = validate_for_template(self.descriptor)
+        descriptor = logic.validate_for_template(self.descriptor)
         self.assertEqual(self.descriptor, descriptor)
 
     def test_validate_for_jinja_modifies_descriptor_if_licenses_is_dict(self):
         self.descriptor['licenses'] = {'url': 'test/url', 'type': 'Test'}
-        descriptor = validate_for_template(self.descriptor)
+        descriptor = logic.validate_for_template(self.descriptor)
         self.assertEqual(descriptor['license'], 'Test')
 
     def test_validate_for_not_errors_if_licenses_is_dict_and_has_no_type_key(self):
         self.descriptor['licenses'] = {'url': 'test/url'}
-        descriptor = validate_for_template(self.descriptor)
+        descriptor = logic.validate_for_template(self.descriptor)
         self.assertEqual(descriptor['license'], None)
 
     def test_validate_for_jinja_removes_licenses_if_invalid_type(self):
         self.descriptor['licenses'] = 1
-        descriptor = validate_for_template(self.descriptor)
+        descriptor = logic.validate_for_template(self.descriptor)
         self.assertEqual(descriptor.get('licenses'), None)
