+++ tests/package/test_controller.py
@@ -12,7 +12,7 @@ from moto import mock_s3
 from app import create_app
 from app.bitstore import BitStore
 from app.database import db
-from app.logic import db_logic
+import app.logic as logic
 from app.package.models import Package, PackageStateEnum, PackageTag
 from app.profile.models import User, Publisher, UserRoleEnum, PublisherUser
 
@@ -95,7 +95,7 @@ class GetMetaDataTestCase(unittest.TestCase):
             publisher.packages.append(metadata)
             db.session.add(publisher)
             db.session.commit()
-            db_logic.Package.delete(self.publisher, self.package)
+            logic.Package.delete(self.publisher, self.package)
         response = self.client. \
             get('/api/package/%s/%s' % (self.publisher, self.package))
         self.assertEqual(response.status_code, 404)
@@ -206,7 +206,7 @@ class FinalizeMetaDataTestCase(unittest.TestCase):
         data = json.loads(response.data)
         self.jwt1 = data['token']
 
-    @patch('app.logic.db_logic.Package.create_or_update')
+    @patch('app.logic.Package.create_or_update')
     @patch('app.bitstore.BitStore.get_metadata_body')
     @patch('app.bitstore.BitStore.get_readme_object_key')
     @patch('app.bitstore.BitStore.get_s3_object')
@@ -296,8 +296,8 @@ class EndToEndTestCase(unittest.TestCase):
             db.session.commit()
 
     @patch('app.bitstore.BitStore.copy_to_new_version')
-    @patch('app.logic.db_logic.Package.create_or_update_tag')
-    @patch('app.logic.db_logic.Package.create_or_update')
+    @patch('app.logic.Package.create_or_update_tag')
+    @patch('app.logic.Package.create_or_update')
     @patch('app.bitstore.BitStore.get_metadata_body')
     @patch('app.bitstore.BitStore.get_readme_object_key')
     @patch('app.bitstore.BitStore.get_s3_object')
@@ -447,7 +447,7 @@ class SoftDeleteTestCase(unittest.TestCase):
         self.jwt_member = data['token']
 
     @patch('app.bitstore.BitStore.change_acl')
-    @patch('app.logic.db_logic.Package.change_status')
+    @patch('app.logic.Package.change_status')
     def test_return_200_if_all_goes_well(self, change_status, change_acl):
         change_acl.return_value = True
         change_status.return_value = True
@@ -456,7 +456,7 @@ class SoftDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 200)
 
     @patch('app.bitstore.BitStore.change_acl')
-    @patch('app.logic.db_logic.Package.change_status')
+    @patch('app.logic.Package.change_status')
     def test_return_403_not_allowed_to_do_operation(self, change_status, change_acl):
         change_acl.return_value = True
         change_status.return_value = True
@@ -466,7 +466,7 @@ class SoftDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 403)
 
     @patch('app.bitstore.BitStore.change_acl')
-    @patch('app.logic.db_logic.Package.change_status')
+    @patch('app.logic.Package.change_status')
     def test_return_401_if_not_header(self, change_status, change_acl):
         change_acl.return_value = True
         change_status.return_value = True
@@ -475,7 +475,7 @@ class SoftDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 401)
 
     @patch('app.bitstore.BitStore.change_acl')
-    @patch('app.logic.db_logic.Package.change_status')
+    @patch('app.logic.Package.change_status')
     def test_throw_500_if_change_acl_fails(self,  change_status, change_acl):
         change_acl.side_effect = Exception('failed')
         change_status.return_value = True
@@ -484,7 +484,7 @@ class SoftDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 500)
 
     @patch('app.bitstore.BitStore.change_acl')
-    @patch('app.logic.db_logic.Package.change_status')
+    @patch('app.logic.Package.change_status')
     def test_throw_500_if_change_status_fails(self, change_status, change_acl):
         change_acl.return_value = True
         change_status.side_effect = Exception('failed')
@@ -493,7 +493,7 @@ class SoftDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 500)
 
     @patch('app.bitstore.BitStore.change_acl')
-    @patch('app.logic.db_logic.Package.change_status')
+    @patch('app.logic.Package.change_status')
     def test_throw_generic_error_if_internal_error(self, change_status, change_acl):
         change_acl.side_effect = Exception('failed')
         change_status.side_effect = Exception('failed')
@@ -569,7 +569,7 @@ class HardDeleteTestCase(unittest.TestCase):
         self.jwt_member = data['token']
 
     @patch('app.bitstore.BitStore.delete_data_package')
-    @patch('app.logic.db_logic.Package.delete')
+    @patch('app.logic.Package.delete')
     def test_return_200_if_all_goes_well(self, db_delete, bitstore_delete):
         bitstore_delete.return_value = True
         db_delete.return_value = True
@@ -578,7 +578,7 @@ class HardDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 200)
 
     @patch('app.bitstore.BitStore.delete_data_package')
-    @patch('app.logic.db_logic.Package.delete')
+    @patch('app.logic.Package.delete')
     def test_throw_500_if_change_acl_fails(self, db_delete, bitstore_delete):
         bitstore_delete.side_effect = Exception('failed')
         db_delete.return_value = True
@@ -588,7 +588,7 @@ class HardDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 500)
 
     @patch('app.bitstore.BitStore.delete_data_package')
-    @patch('app.logic.db_logic.Package.delete')
+    @patch('app.logic.Package.delete')
     def test_throw_500_if_change_status_fails(self, db_delete, bitstore_delete):
         bitstore_delete.return_value = True
         db_delete.side_effect = Exception('failed')
@@ -598,7 +598,7 @@ class HardDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 500)
 
     @patch('app.bitstore.BitStore.delete_data_package')
-    @patch('app.logic.db_logic.Package.delete')
+    @patch('app.logic.Package.delete')
     def test_throw_generic_error_if_internal_error(self, db_delete, bitstore_delete):
         bitstore_delete.side_effect = Exception('failed')
         db_delete.return_value = False
@@ -608,7 +608,7 @@ class HardDeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 500)
 
     @patch('app.bitstore.BitStore.delete_data_package')
-    @patch('app.logic.db_logic.Package.delete')
+    @patch('app.logic.Package.delete')
     def test_should_throw_403_if_user_is_not_owner_of_the_package(self,
                                                                   db_delete,
                                                                   bitstore_delete):
@@ -704,7 +704,7 @@ class UndeleteTestCase(unittest.TestCase):
         self.jwt_non_member = data['token']
 
     @patch('app.bitstore.BitStore.change_acl')
-    @patch('app.logic.db_logic.Package.change_status')
+    @patch('app.logic.Package.change_status')
     def test_should_return_200_if_all_goes_well(self, change_status,
                                                 change_acl):
         change_acl.return_value = True
@@ -714,7 +714,7 @@ class UndeleteTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 200)
 
     @patch('app.bitstore.BitStore.change_acl')
-    @patch('app.logic.db_logic.Package.change_status')
+    @patch('app.logic.Package.change_status')
     def test_return_403_not_allowed_to_do_operation(self, change_status, change_acl):
         auth = "%s" % self.jwt_non_member
         response = self.client.post(self.url, headers={'Auth-Token': auth})
@@ -817,7 +817,7 @@ class TagDataPackageTestCase(unittest.TestCase):
         self.auth = "%s" % self.jwt
 
     @patch('app.bitstore.BitStore.copy_to_new_version')
-    @patch('app.logic.db_logic.Package.create_or_update_tag')
+    @patch('app.logic.Package.create_or_update_tag')
     def test_return_200_if_all_goes_well(self, create_or_update_tag,
                                          copy_to_new_version):
         copy_to_new_version.return_value = True
@@ -831,7 +831,7 @@ class TagDataPackageTestCase(unittest.TestCase):
         self.assertEqual(response.status_code, 200)
 
     @patch('app.bitstore.BitStore.copy_to_new_version')
-    @patch('app.logic.db_logic.Package.create_or_update_tag')
+    @patch('app.logic.Package.create_or_update_tag')
     def test_throw_400_if_version_missing(self, create_or_update_tag,
                                           copy_to_new_version):
         copy_to_new_version.return_value = True
@@ -847,7 +847,7 @@ class TagDataPackageTestCase(unittest.TestCase):
         self.assertEqual('version not found', data['message'])
 
     @patch('app.bitstore.BitStore.copy_to_new_version')
-    @patch('app.logic.db_logic.Package.create_or_update_tag')
+    @patch('app.logic.Package.create_or_update_tag')
     def test_throw_500_if_failed_to_tag(self, create_or_update_tag,
                                         copy_to_new_version):
         copy_to_new_version.side_effect = Exception('failed')
@@ -903,7 +903,7 @@ class TagDataPackageTestCase(unittest.TestCase):
         self.assertTrue('Contents' not in objects_nu)
 
     @patch('app.bitstore.BitStore.copy_to_new_version')
-    @patch('app.logic.db_logic.Package.create_or_update_tag')
+    @patch('app.logic.Package.create_or_update_tag')
     def test_allow_if_member_of_publisher(self, create_or_update_tag,
                                           copy_to_new_version):
         copy_to_new_version.side_effect = Exception('failed')
